Atari 7800 - sort out system id stuff
