+++ BizHawk.Client.ApiHawk/Classes/BizHawkSystemIdToCoreSystemEnumConverter.cs
@@ -32,9 +32,6 @@ namespace BizHawk.Client.ApiHawk
 					return CoreSystem.Atari2600;
 
 				case "A78":
-					return CoreSystem.Atari2600;
-
-				case "A7800":
 					return CoreSystem.Atari7800;
 
 				case "Coleco":
