+++ BizHawk.Client.EmuHawk/BizHawk.Client.EmuHawk.csproj
@@ -1167,7 +1167,6 @@
     </Compile>
     <Compile Include="tools\VirtualPads\schema\A26Schema.cs" />
     <Compile Include="tools\VirtualPads\schema\A78Schema.cs" />
-    <Compile Include="tools\VirtualPads\schema\A7800HawkSchema.cs" />
     <Compile Include="tools\VirtualPads\schema\AppleIISchema.cs" />
     <Compile Include="tools\VirtualPads\schema\C64Schema.cs" />
     <Compile Include="tools\VirtualPads\schema\ColecoSchema.cs" />
@@ -1317,7 +1316,7 @@
     <EmbeddedResource Include="config\A7800\A7800ControllerSettings.resx">
       <DependentUpon>A7800ControllerSettings.cs</DependentUpon>
     </EmbeddedResource>
-		<EmbeddedResource Include="config\A7800\A7800FilterSettings.resx">
+    <EmbeddedResource Include="config\A7800\A7800FilterSettings.resx">
       <DependentUpon>A7800FilterSettings.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="config\INTV\IntvControllerSettings.resx">
